+++ src/main/java/com/deliverable/repositories/TicketRepository.java
@@ -5,9 +5,9 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.deliverable.model.Ticket;
 
-public interface TicketRepository extends JpaRepository<Ticket, Integer>, TicketRepositoryBase {
+public interface TicketRepository extends JpaRepository<Ticket, Long>, TicketRepositoryBase {
 	
-	public Ticket findTicketById(Integer id);
+	public Ticket findTicketById(Long id);
 	
 	public List<Ticket> findTicketByStatusValueNotOrderByPriorityWeightDescDateCreated(String statusValue);
