+++ src/main/java/com/deliverable/repositories/TicketRepositoryBase.java
@@ -6,8 +6,8 @@ import com.deliverable.model.Transition;
 
 public interface TicketRepositoryBase {
 
-	public void updateTicketStatus(Integer ticketId, Integer statusId);
+	public void updateTicketStatus(Long ticketId, Long statusId);
 	
-	public List<Transition> getTransitions(Integer ticketTypeId, Integer originStatusId);
+	public List<Transition> getTransitions(Long ticketTypeId, Long originStatusId);
 	
 }
